pyproject.toml
[project]
name = "django-docker-container"
version = "0.1.0"
description = "Django application running in a Docker container"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asgiref==3.9.1",
    "django==5.2.4",
    "django-allauth==65.9.0",
    "django-browser-reload==1.18.0",
    "django-cleanup==9.0.0",
    "django-htmx==1.23.2",
    "django-tinymce>=4.1.0",
    "markdown==3.8.2",
    "pillow==11.3.0",
    "python-dotenv>=1.1.1",
    "sqlparse==0.5.3",
    "tzdata==2025.2",
]
